requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
# TensorFlow and scikit-learn are required for the improved training workflow.
tensorflow>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.3
pillow>=10.0.0
python-multipart==0.0.6
aiofiles>=23.2.1
pydantic>=2.4.2
typing-extensions>=4.5.0
